requirements.txt
discord.py
python-dotenv
requests
pandas
thefuzz
flask
waitress